feat: create KafkaSourceCallback class
