+++ bizon/source/source.py
@@ -3,10 +3,10 @@ from typing import Any, List, Optional, Tuple, Type, Union
 
 from requests.auth import AuthBase
 
+from .callback import AbstractSourceCallback, NoOpSourceCallback
 from .config import SourceConfig
 from .models import SourceIncrementalState, SourceIteration
 from .session import Session
-from .callback import AbstractSourceCallback, NoOpSourceCallback
 
 
 class AbstractSource(ABC):
